% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectionFunctions.R
\name{shutdownClient}
\alias{shutdownClient}
\title{Extend the visibility of the shutdownClient function in J4R.}
\usage{
shutdownClient()
}
\description{
Extend the visibility of the shutdownClient function in J4R.
}
